llm_api_on_ec2/requirements.txt
google-genai
pydantic
httpx